Move `ScheduleCallable` to `thinc.types`
